refactor: 패키지 구조 변경
DDD에 맞춘 global 패키지 추가 및 수정
