--- src/main/java/apptive/team1/friendly/common/jwt/JwtTokenProvider.java
@@ -1,105 +1,0 @@
-package apptive.team1.friendly.common.jwt;
-
-import io.jsonwebtoken.*;
-import io.jsonwebtoken.io.Decoders;
-import io.jsonwebtoken.security.Keys;
-import org.springframework.security.authentication.UsernamePasswordAuthenticationToken;
-import org.springframework.security.core.Authentication;
-import org.springframework.security.core.GrantedAuthority;
-import org.springframework.security.core.authority.SimpleGrantedAuthority;
-import org.springframework.security.core.userdetails.User;
-
-import java.security.Key;
-import java.util.Arrays;
-import java.util.Collection;
-import java.util.Date;
-import java.util.stream.Collectors;
-
-public class JwtTokenProvider {
-
-    protected static final String AUTHORITIES_KEY = "auth";
-
-    protected final String secret;
-    protected final long tokenValidityInMilliseconds;
-
-    protected Key key;
-
-    public JwtTokenProvider(String secret, long tokenValidityInMilliseconds) {
-        this.secret = secret;
-        this.tokenValidityInMilliseconds = tokenValidityInMilliseconds * 1000;
-
-        byte[] keyBytes = Decoders.BASE64.decode(secret);
-        this.key = Keys.hmacShaKeyFor(keyBytes);
-    }
-
-    /**
-     * 토큰 생성
-     */
-    public String createToken(Authentication authentication) {
-        String authorities = authentication.getAuthorities()
-                .stream()
-                .map(GrantedAuthority::getAuthority)
-                .collect(Collectors.joining(","));
-
-        long now = (new Date()).getTime();
-        Date validity = new Date(now + this.tokenValidityInMilliseconds);
-
-        return Jwts.builder()
-                .setSubject(authentication.getName())
-                .claim(AUTHORITIES_KEY, authorities)
-                .signWith(key, SignatureAlgorithm.HS512)
-                .setExpiration(validity)
-                .compact();
-    }
-
-    /**
-     * token -> claims
-     */
-    public Claims getClaimsFromToken(String token) {
-        return Jwts.parserBuilder()
-                .setSigningKey(key)
-                .build()
-                .parseClaimsJws(token)
-                .getBody();
-    }
-
-    /**
-     * token -> Authentication
-     */
-    public Authentication getAuthentication(String token) {
-        // token -> claims
-        Claims claims = getClaimsFromToken(token);
-
-        // claims -> authorities
-        Collection<? extends GrantedAuthority> authorities
-                = Arrays.stream(claims.get(AUTHORITIES_KEY).toString().split(","))
-                .map(SimpleGrantedAuthority::new)
-                .collect(Collectors.toList());
-
-        // claims, authorities -> User 객체
-       User principal = new User(claims.getSubject(), "", authorities);
-
-       // 최종적으로 Authentication 객체를 return
-        return new UsernamePasswordAuthenticationToken(principal, token, authorities);
-    }
-
-    /**
-     * 토큰 유효성 검사
-     */
-    // TODO logger 설정 추가후에 System.out.println을 looger.info로 바꿔줄 것
-    public boolean validateToken(String token) {
-        try {
-            Jwts.parserBuilder().setSigningKey(key).build().parseClaimsJws(token);
-            return true;
-        } catch (io.jsonwebtoken.security.SecurityException | MalformedJwtException e) {
-            System.out.println("잘못된 JWT 서명입니다.");
-        } catch (ExpiredJwtException e) {
-            System.out.println("만료된 JWT 토큰입니다.");
-        } catch (UnsupportedJwtException e) {
-            System.out.println("지원되지 않는 JWT 토큰입니다.");
-        } catch (IllegalArgumentException e) {
-            System.out.println("잘못된 JWT 토큰입니다.");
-        }
-        return false;
-    }
-}
