--- src/main/java/apptive/team1/friendly/common/jwt/config/JwtSecurityConfig.java
@@ -1,23 +1,0 @@
-package apptive.team1.friendly.common.jwt.config;
-
-import apptive.team1.friendly.common.jwt.JwtFilter;
-import apptive.team1.friendly.common.jwt.JwtTokenProvider;
-import org.springframework.security.config.annotation.SecurityConfigurerAdapter;
-import org.springframework.security.config.annotation.web.builders.HttpSecurity;
-import org.springframework.security.web.DefaultSecurityFilterChain;
-import org.springframework.security.web.authentication.UsernamePasswordAuthenticationFilter;
-
-public class JwtSecurityConfig extends SecurityConfigurerAdapter<DefaultSecurityFilterChain, HttpSecurity> {
-
-    private JwtTokenProvider jwtTokenProvider;
-
-    public JwtSecurityConfig(JwtTokenProvider jwtTokenProvider) {
-        this.jwtTokenProvider = jwtTokenProvider;
-    }
-
-    @Override
-    public void configure(HttpSecurity httpSecurity) {
-        JwtFilter customFilter = new JwtFilter(jwtTokenProvider);
-        httpSecurity.addFilterBefore(customFilter, UsernamePasswordAuthenticationFilter.class);
-    }
-}
