+++ src/main/java/apptive/team1/friendly/domain/user/controller/UserController.java
@@ -1,8 +1,9 @@
 package apptive.team1.friendly.domain.user.controller;
 
+import apptive.team1.friendly.domain.user.data.dto.AccountInfoResponse;
+import apptive.team1.friendly.domain.user.data.dto.GoogleSignUpRequest;
 import apptive.team1.friendly.domain.user.data.dto.SignupRequest;
 import apptive.team1.friendly.domain.user.data.dto.SignupResponse;
-import apptive.team1.friendly.domain.user.data.dto.AccountInfoResponse;
 import apptive.team1.friendly.domain.user.data.dto.profile.ProfileImgDto;
 import apptive.team1.friendly.domain.user.service.UserService;
 import org.springframework.http.HttpStatus;
