--- src/main/java/apptive/team1/friendly/domain/user/data/dto/GoogleLoginResponse.java
@@ -1,16 +1,0 @@
-package apptive.team1.friendly.domain.user.data.dto;
-
-import lombok.AllArgsConstructor;
-import lombok.Builder;
-import lombok.Data;
-import lombok.NoArgsConstructor;
-
-@Data
-@Builder
-@NoArgsConstructor
-@AllArgsConstructor
-public class GoogleLoginResponse {
-
-    private String accessToken;
-    private boolean isRegistered;
-}
