--- src/main/java/apptive/team1/friendly/domain/user/data/dto/LoginRequest.java
@@ -1,22 +1,0 @@
-package apptive.team1.friendly.domain.user.data.dto;
-
-import lombok.*;
-
-import javax.validation.constraints.NotNull;
-import javax.validation.constraints.Size;
-
-@Getter
-@Setter
-@Builder
-@AllArgsConstructor
-@NoArgsConstructor
-public class LoginRequest {
-
-    @NotNull
-    @Size(min = 3, max = 50)
-    private String email;
-
-    @NotNull
-    @Size(min = 3, max = 50)
-    private String password;
-}
