--- src/main/java/apptive/team1/friendly/domain/user/data/dto/LoginResponse.java
@@ -1,13 +1,0 @@
-package apptive.team1.friendly.domain.user.data.dto;
-
-import lombok.*;
-
-@Getter
-@Setter
-@Builder
-@NoArgsConstructor
-@AllArgsConstructor
-public class LoginResponse {
-
-    private String accessToken;
-}
