--- src/main/java/apptive/team1/friendly/domain/user/service/AuthService.java
@@ -1,39 +1,0 @@
-package apptive.team1.friendly.domain.user.service;
-
-import apptive.team1.friendly.domain.user.data.dto.LoginResponse;
-import apptive.team1.friendly.domain.user.data.repository.AccountRepository;
-import apptive.team1.friendly.common.jwt.JwtTokenProvider;
-import org.springframework.security.authentication.UsernamePasswordAuthenticationToken;
-import org.springframework.security.config.annotation.authentication.builders.AuthenticationManagerBuilder;
-import org.springframework.security.core.Authentication;
-import org.springframework.security.core.context.SecurityContextHolder;
-import org.springframework.stereotype.Service;
-
-@Service
-public class AuthService {
-
-    private final JwtTokenProvider jwtTokenProvider;
-    private final AuthenticationManagerBuilder authenticationManagerBuilder;
-    private final AccountRepository accountRepository;
-
-    public AuthService(JwtTokenProvider jwtTokenProvider, AuthenticationManagerBuilder authenticationManagerBuilder, AccountRepository accountRepository) {
-        this.jwtTokenProvider = jwtTokenProvider;
-        this.authenticationManagerBuilder = authenticationManagerBuilder;
-        this.accountRepository = accountRepository;
-    }
-
-    public LoginResponse authenticate(String username, String password) {
-        UsernamePasswordAuthenticationToken authenticationToken =
-                new UsernamePasswordAuthenticationToken(username, password);
-
-        // 이 과정에서 CustomUserDetailsService 에서 우리가 재정의한 loadUserByUsername 메서드 호출
-        Authentication authentication = authenticationManagerBuilder.getObject().authenticate(authenticationToken);
-        SecurityContextHolder.getContext().setAuthentication(authentication);
-
-        String jwt = jwtTokenProvider.createToken(authentication);
-
-        return LoginResponse.builder()
-                .accessToken(jwt)
-                .build();
-    }
-}
