+++ src/main/java/apptive/team1/friendly/global/auth/local/dto/LoginRequest.java
@@ -1,0 +1,22 @@
+package apptive.team1.friendly.global.auth.local.dto;
+
+import lombok.*;
+
+import javax.validation.constraints.NotNull;
+import javax.validation.constraints.Size;
+
+@Getter
+@Setter
+@Builder
+@AllArgsConstructor
+@NoArgsConstructor
+public class LoginRequest {
+
+    @NotNull
+    @Size(min = 3, max = 50)
+    private String email;
+
+    @NotNull
+    @Size(min = 3, max = 50)
+    private String password;
+}
