+++ src/main/java/apptive/team1/friendly/global/auth/local/dto/LoginResponse.java
@@ -1,0 +1,13 @@
+package apptive.team1.friendly.global.auth.local.dto;
+
+import lombok.*;
+
+@Getter
+@Setter
+@Builder
+@NoArgsConstructor
+@AllArgsConstructor
+public class LoginResponse {
+
+    private String accessToken;
+}
