+++ src/main/java/apptive/team1/friendly/global/auth/social/dto/SocialAuthResponse.java
@@ -1,0 +1,16 @@
+package apptive.team1.friendly.global.auth.social.dto;
+
+import lombok.AllArgsConstructor;
+import lombok.Builder;
+import lombok.Data;
+import lombok.NoArgsConstructor;
+
+@Data
+@Builder
+@NoArgsConstructor
+@AllArgsConstructor
+public class GoogleLoginResponse {
+
+    private String accessToken;
+    private boolean isRegistered;
+}
